+++ src/devices/colorbulb.ts
@@ -215,11 +215,18 @@ export class ColorBulb {
   }
 
   config(device: device & devicesConfig) {
-    if (device.bot !== undefined) {
-      device.bot['ble'] = device.ble;
-      device.bot['logging'] = this.deviceLogging;
-      device.bot['refreshRate'] = this.deviceRefreshRate;
-      this.warnLog(`Color Bulb: ${this.accessory.displayName} Config: ${JSON.stringify(device.colorbulb)}`);
+    const config: any = device.colorbulb;
+    if (device.colorbulb !== undefined) {
+      if (device.ble !== undefined) {
+        config['ble'] = device.ble;
+      }
+      if (device.logging !== undefined) {
+        config['logging'] = device.logging;
+      }
+      if (device.refreshRate !== undefined) {
+        config['refreshRate'] = device.refreshRate;
+      }
+      this.warnLog(`Color Bulb: ${this.accessory.displayName} Config: ${JSON.stringify(config)}`);
     }
   }